requirements.txt
decorator==4.0.10
Django==1.9.4
django-extensions==1.7.4
django-haystack==2.5.0
elasticsearch==2.4.0
ipython==5.1.0
ipython-genutils==0.1.0
model-mommy==1.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.0
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.0
urllib3==1.17
wcwidth==0.1.7

django-ckeditor==5.1.1